Refactor to packages
To make things maintainable we have split logic to few packages.

"app" package contains domain types without any dependencies -
currently Article. It also defines an *interface* for working with
articles - ArticlesService. We have one implementation of this service
backed by Postgres and it's in postgres package. Everything related to
database, SQL and Postgres db is in this package.

Server type now glues everything together - it creates concrete
ArticlesService (postgres-based) and uses it in handlers.

All of this gives us a lot of good things:

1. You can get a sense of what this application is about by looking at
   types and interfaces in "app" package.
2. HTTP handlers are very simple - they just invoke methods of
   ArticlesService and wrap them in HTTP.
3. Database logic is located in particular package.

This approach was stolen from/inspired by this post -
https://medium.com/@benbjohnson/standard-package-layout-7cdbc8391fc1