Rewrite handlers using gorilla/mux instead of gin
During testing I've found it cumbersome to test gin handlers as they are
not comforming to http.Handler. In particular it's not possible to test
for 404 code. Also, writing handlers using gin makes them unsuitable for
using in httputil.ResponseRecorder and unsuitable to the vast ecosystem
of standard tools.

So instead we are using gorilla/mux for routing. We don't use standard
ServeMux because we want convenient way to pass request arguments, e.g.
/articles/{slug} to the handler.

This change rewrites handlers as http.HandlerFunc for gorilla/mux. We
also rewrite tests to check code and body.

The downside for now is that handlers now a little bit more verbose
because they do explicit JSON marshalling and writing headers and test
code is more verbose. We'll see how it will play with more handlers.