Rewrite server tests to make them integrational
Server tests should test handlers and need a routing fully configured so
handlers would get parameters from URL. This allow us to test the actual
behavior of handlers - how they handle non-existing items, whether they
return correct status code, etc.

To implement this we setup a server with mock services in TestMain by
supplying config to the NewServer method. The created server is stored
as a package variable and used in all test functions.

For article handlers we've rewritten test to be table-driven and test 3
cases:

1. Get articles list, check 200 code and body for substring.
2. Get non-existing article,  check 404 code.
3. Get existing article, check 200 code and body for substring.