Accept UserRequest in UserService.Register
Up until now Register accepted User object and returned just error. But
the response object has to contain User object so Register handler just
reused the user object from client request -- this was strange as we're
returning to the client its data.

Now Register accepts UserRequest which is more semantically right and
it's also returns User object as it's stored in the service.

Again, this is all about making it more correct from semantic point of
view.