Add errors values for user exists and not found
Previosuly, handlers returned internal server error (500) for cases when
we tried to register already existing user or login with a user that is
not existing. These cases are not server errors (5xx), they are client
errors but couldn't distinguish these errors from user service.

This commit adds 2 error values - app.ErrUserExists and
app.ErrUserNotFound that are returned in case of these errors. In
handlers we check for these errors and return correct client error.