Run app, db and migrations in docker-compose
To quickly iterate we need to run the whole stack. This is where
docker-compose shines. Previously, only db was run in docker-compose,
though it was run in host mode.

The idea of these changes is to run everythin in docker-compose isolated
in its own network. So now we have 3 containers:

1. db that runs Postgres. Database and user are automatically created by
   passing POSTGRES_DB and POSTGRES_USER env vars (this is a feauture of
   the official image).
2. migrate that applies migrations. Migration image is custom built from
   migrate/migrate by adding "wait-for" script that simply waits when
   Postgres will open port 5432 and be ready to accept requests and then
   applies migrations.
3. app that runs conduit backend. It's built with multi-stage build and
   it's the only container that exposes port to the host.