Remove sample data migration
We had a migration for inserting sample data - I thought that it would
be useful for test. But now I think that migrations should contain only
schema without any data because otherwise we can forget about it and
deploy sample users on prod.

Instead of sample we will create users for testing in setup code
(`TestMain`). This way it will be more clear because test data will be
in the same file with test code.