Use vendoring
In container for integration testing `go test` was always downloading
deps because container doesn't have them. To make vendoring work with
modules we have to pass `-mod=vendor` to all of the go commands. The
most convenient way for this is via GOFLAGS env variable introduced in
go 1.11.

This commit adds use of vendor dir in docker image for integration tests
and into Makefile.