Make email primary key in users table
It's a good practice to avoid syntetic columns as a primary keys. And
since email is unique why not make it a primary key - searching is
performing using email anyway. And id is not needed anymore.

This commit modifies migration to use email as a primary key in users
table.