Add migration to create followers table
To store followers relations we use classic many-to-many reference
table. "follower" fields is a follower user, "follows" is a user that
follower follows. Since one row describe a single relationship and user
may follow many users there will be duplicated follower fields. So we
can't make follower a primary key. But because foreign keys are
references primary keys checks on whether a referencing row has a match
will be efficient.