Add user following feature
Following a user is a simple record of a follower user and the other
user that is followed. This record is stored in a "followers" table
which is a usual many-to-many relation. It has 2 fields - both are
referencing name field in the user table. The reason why it references
name and not the email, which is a primary key, is that according to API
it works with so-called "profiles" that are identified by username. The
relation in "followers" table has a unique constraint with both fields
to avoid duplicates and it also allows efficient query to check whether
one user follows another because unique constraint is automatically
backed by btree index in Postgres.

To make a follow request user makes authenticated POST request to the
/profiles/{username}/follow endpoint. This will create a following
record with the current user from JWT as a follower and username in URL
as a user to follow.

This commit adds 4 things:

1. Update migration to create a working schema for the following
   feature - referencing name, unique constraint.
2. User service method that create record in the DB table. It uses
   UPSERT feature of Postgres (ON CONFLICT DO NOTHING) to gracefully
   handle duplicated records.
3. API handler. Nothing fancy here.
4. Following support to CLI.