Fix gitignore for binaries
We had a rule in gitignore to ignore "conduit" and "hashpass". My search
tool "ag" respects gitignore but it was interpreting it as ignoring
these strings anywhere in the path. Because of this I couldn't open
these files in Vim.

This commit fixes gitignore rule to correctly ignore binaries in the top
level dir.