jwt: Store username in token claims
Username is needed for articles feed handlers to quickly get current
user name and avoid additional query to users db.

This commit adds changes signature of New method to accept *app.User
struct and fixes usage of it in user service. It also adds check of
username in Auth middleware and fixes tests in user service.