Add articles feed feature
Users may list global feed of articles via GET /articles/ endpoint. But
because they can follow other users they want to receive feed of
articles of users they follow. This is the implementation of this
feature.

Articles feed can be received via GET /articles/feed endpoint. This
endpoint is authenticated because it requires knowledge of the current
user. It uses "username" claim from JWT added in previous commits.

In the databassse Articles feed is queried pretty much the same as
global list with the additional filter on follower field equals to
current username.